Updated Octree Class to further reduce file size to ~0.6 of previous size, because every sub octree was saving a position even though this is redundant. I don't think I can get the data smaller without changing the stored datatype or using zlib for compression.

+++ source/include/world/chunk.h
@@ -6,7 +6,7 @@
 class Chunk{
 public:
   //Position information and size information
-  glm::vec3 pos;
+  glm::ivec3 pos;
   bool remesh = false;
   int size = 16;
   BiomeType biome;
